requirements.txt
fastapi==0.73.0
starlette_exporter==0.12.0
# NOTE: need to downgrade starlette due to fastapi dependency
starlette==0.17.1
pydantic==1.9.0
requests==2.27.1
pygithub==1.55
python-slugify==5.0.2

gunicorn==20.1.0
uvicorn[standard]==0.17.2

pytest==6.2.5
pytest-watch==4.2.0
flake8==4.0.1
mypy
black
ipdb
types-python-slugify
